requirements.txt
# Core ML Libraries
torch>=2.1.0
transformers>=4.36.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Data Processing
datasets>=2.15.0
pandas>=2.1.0
numpy>=1.24.0

# OpenAI API
openai>=1.6.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
scipy>=1.11.0

# Logging & Visualization
tensorboard>=2.15.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Development
jupyter>=1.0.0
ipywidgets>=8.1.0
ruff>=0.1.0
